pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cortex-pkm"
version = "0.1.4"
description = "Plain text knowledge management"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "lbugnon", email = "lbugnon@sinc.unl.edu.ar"}
]
keywords = ["markdown", "knowledge-management", "notes", "pkm", "productivity"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
    "Topic :: Text Processing :: Markup :: Markdown",
]
dependencies = [
    "click>=8.0",
    "python-frontmatter>=1.0",
    "rapidfuzz>=3.0",
    "simple-term-menu>=1.6",
    "habanero>=1.2",
    "arxiv>=2.0.0",
    "bibtexparser>=1.4",
    "dateparser>=1.2",
]

[project.urls]
Repository = "https://github.com/lbugnon/cortex_pkm"

[project.scripts]
cor = "cor.cli:cli"

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["cor*"]

[tool.setuptools.package-data]
cor = ["assets/*.md", "assets/*.yaml", "hooks/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
